requirements.txt
fastapi
uvicorn
sqlmodel
SQLAlchemy
passlib[bcrypt]
pyjwt
python-jose[cryptography]